Move parse for opencontrol to root of package
it reads better now as `opencontrol.Parse`
