+++ commands/get/get.go
@@ -2,9 +2,9 @@ package get
 
 import (
 	"github.com/opencontrol/compliance-masonry/lib/common"
-	"github.com/opencontrol/compliance-masonry/lib/opencontrol/versions"
 	"github.com/opencontrol/compliance-masonry/lib/opencontrol/versions/base"
 	"github.com/opencontrol/compliance-masonry/lib/opencontrol/parser"
+	"github.com/opencontrol/compliance-masonry/lib/opencontrol"
 )
 
 // Get will retrieve all of the resources for the schemas and the resources for all the dependent schemas.
@@ -14,7 +14,7 @@ func Get(destination string, configData []byte, worker *base.Worker) error {
 		return common.ErrNoDataToParse
 	}
 	// Parse it.
-	configSchema, err := versions.Parse(parser.Parser{}, configData)
+	configSchema, err := opencontrol.Parse(parser.Parser{}, configData)
 	if err != nil {
 		return err
 	}
