+++ lib/opencontrol/parse.go
@@ -1,4 +1,4 @@
-package versions
+package opencontrol
 
 import (
 	"errors"
